"Fix" things in code I stole from phpish/link_header
Not sure if this works, but it seems to me, that a lot of these
if-statements could be condensed. Also PHP was giving me this warning:

PHP Warning: "continue" targeting switch is equivalent to "break".
Did you mean to use "continue 2"?

StackOverflow says "continue 2" and something about PHP 7.3, I'm not
sure what's the intended behaviour here, because I suck at writing
and understanding parsers.

For now, I just removed the "continue" to get rid of the warnings. It
shouldn't change anything according to the PHP docs, since 'break' and
'continue' appparently have identical behaviour in foreach loops,
but maybe it was already broken?

Idk.