Improve security
I disabled direct access to any PHP files, because that could be
a pretty big security risk. Now, all traffic is routed to index.php,
which does the routing.

I made an exception for the files in /endpoint, because those are meant
to be accessed directly. I also made the accessible without the .php
suffix.